misc: moved set_config() and update_version() to utils.py
